discover: new config parser
This patch implements a different config parser for the go-discover
config language. This is a breaking change from the previous
implementation since the values are no longer expected to be URL encoded
but are taken as literals unless they contain a space, backslash or
double-quote. In that case they need to be quoted and special characters
can be escaped as usual.