tests: remove test helpers
In an effort to automate the testing of go-disover, this
removes all but the Terraform configuration necessary to create
target resources to filter based on API requests made by `go test`.

This is a different approach than prior, and does not require binary
upload to the hosts. This removes our ability to test metadata APIs
that are only avaiable from the host. However, go-discover
performs very little logic specific to those metadata APIs. Largely,
it is passing the environment through to the client SDKs. With
just credential authentication to internet-accessible APIs, we can
get a fair amount of confidence, and the simplified testing
story will let us run the tests more readily, including in CI.