provider/vsphere: Add vSphere discovery
This adds vSphere discovery based on tags. Discovery is done by locating
the ID of a tag based on its name, and the name of the category it
belongs to. All IP addresses for matching VMs are reported (save
loopback and auto-config addresses), so long as VMware tools is running
on each matching guest.

This provider requires vCenter, and at least vSphere 6.0 to properly
support the tagging API REST endpoints. It's also recommended, given
that these credentials will ultimately be in the clear, that reduced
access (at the very least, general read-only) be given to the account
used to perform discovery.