provider/vsphere: Don't allow environment for non-credential settings
To stay consistent with the rest of the providers, we are only allowing
credential settings to be set by environment variables. This matches the
behavior of most other SDKs, which allow credentials to be provided
out-of-band via some sort of method, typically the environment. The
vSphere SDK lacks this feature.