openstack: don't look up VMs by project ID.
There is no need to specify the project ID to find VMs. By default you
only see the VMs of the project that you authenticated with anyway.
(And there was no way in the discover code to auth with one project and
look up another.) Finding the ID is annoying if you want to auth by
project name, which was not even possible before.
Since we don't *need* to obtain the ID, we can get rid of the somewhat
ususal way to obtain it from the metadata service.